requirements.txt
pysnyk==0.2.0
PyGithub==1.44.1
slack-webhook==1.0.2